Move checher.rb to /checkers
