+++ lib/npmdc/checkers/checker.rb
